requirements.txt
azure-ai-projects>=1.0.0
azure-core>=1.30.0
azure-search-documents>=11.4.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
requests>=2.31.0
streamlit>=1.35.0
